Updating CI Process.
It's a small team, so no need to go crazy with this for now. If we end up with more than a handful of people on this, we will want to refactor the CI to use tagged branches and proper release branch integration for staging, etc.  But for now, thats just a lot more work to manage than needed for like 2-3 people to maintain this project.