Fixes #3: Update Sheriff Map
Previous code was in place that skipped South Dakota.  We had an incorrect FIPS code in our CSV file that was breaking the entire map, so MapBox worked around it by ignoring the state alltogether.

The CSV is updated, and code changes should prevent future issues where this kind of error can occur.